Reuse text and geometry related headless mocks in Avalonia unit tests and benchmarks
